Various updates
  - Put methods in corresponding mealy filenames
  - Add requirements.txt again
  - Add removed code that is used in the notebok
