+++ code-env/python/spec/requirements.txt
@@ -1,0 +1,11 @@
+graphviz==0.16
+matplotlib==3.3.4
+
+scikit-learn>=0.20,<1.1
+scipy>=1.2,<1.3
+xgboost==0.82
+lightgbm>=3.2,<3.3
+statsmodels>=0.10,<0.11
+jinja2>=2.10,<2.11
+flask>=1.0,<1.1
+cloudpickle>=1.3,<1.6
