+++ python-lib/dku_error_analysis_decision_tree/node.py
@@ -1,5 +1,7 @@
 from math import isnan
-from dku_error_analysis_utils import safe_str, format_float
+
+from dku_error_analysis_utils import safe_str
+from mealy_local.error_analysis_utils import format_float
 from mealy_local.error_analyzer_constants import ErrorAnalyzerConstants
