+++ tests/python/unit/requirements.txt
@@ -2,5 +2,4 @@ pandas~=1.0
 pytest~=6.2
 allure-pytest==2.8.29
 numpy==1.16.6
-mealy==0.2.3
 pytest-mock~=3.6
